Modify web_config resource for dynamic loading
The web_config provider will now render a template that optimistically
loads a named template, default local_settings_dynamic.py, allowing the
inclusion of arbitrary python statements in the graphite web config.

Includes ChefSpec tests for both resource and provider.

Note: an odd rspec bug was encountered in testing the provider code and
required a workaround to eager load the metadata description.

Free HW stickers and an iron on patch for whoever fixes this in rspec!