Remove extraneous re-addition of package list to package list.
To match the pattern used below in the Red Hat portion of the case
statement. += already takes care of including the original list of
Debian packages.

The repetition has been safe to repeat—thanks Chef package resource
idempotency—but should probably be removed to save Chef trying to
reinstall the original list of packages again.

Caught while walking through the code with @LuqiPan