Update indexing script for metrics and runit for uwsgi
The Ubuntu packages leave the script for building metric indexes broken, where
the script doesn't account for graphite customization changing paths, and also
doesn't account for graphite-web having a hardcoded path for the expected
location of the script.

This change makes it so the cookbook updates the script Ubuntu uses to include
the configured location and users. It will also symlink the script from the
location that graphite-web expects it to be in the coded locations. This fixes
both graphite-web manually indexing as well as the cronjob buidling indexes.

This also turns off restart_on_update for runit on uwsgi. On creation, it
tended to be problematic as the enable action started it then restarted. It
seemed uwsgi would miss this signal occassionally and it lead to random
failures on setup.