Increase the restart timeout for carbon cache
We have slowed down the disk writes for carbon-cache-a on some clusters.  This
causes the restarts of carbon cache to take longer then the 7 second timeout
currently set, which is the default timeout for runit.  This change sets the
timeout from the default 7 seconds to 60 seconds for the carbon caches.

Resolves: SRE-399