Increase the time allowed to restart uwsgi
It is possible that uwsgi is taking longer to restart because
of either the new deb or performance in the testing env.  This
might be causing the tests to fail.  So, we are going to test
changing the timeout to 120 seconds to see if the intermittent
failures in the test env go away.

See also: SRE-466